Combine full day/month names with abbreviations
During parsing, we make no distinction between abbreviations and full
names, so the locale data can be a little simpler if it just records
"months" and "weekdays" instead of splitting them into two categories
that we just have to merge later.

Also there are a few locales where the "abbreviated" names are just
equal to the full-length names, so by combining them we get a slightly
smaller `locales/glibc.json`.