Treat prefix/suffix as hints to maximize (very slow!)
Previously, if a prefix or suffix indicated that a particular format
might be a good idea, we'd only consider such formats, except if that
ruled out all legal formats. Maybe. Actually I'm not entirely sure what
the behavior was in all circumstances; it was hard to reason about.

This commit treats our available information in two different ways.

Keywords generated by conversion specifiers still impose hard
requirements on locales, and we still score each possible pattern and
return the patterns which optimize the score.

Prefixes and suffixes, however, are now just treated as hints. The
pattern's score is now the number of hints satisfied plus the number of
conversions.

In addition, a hint now only counts if the corresponding neighbor is
being treated as literal text and not as a conversion. Sometimes we have
to choose between satisfying a hint or treating the neighbor as a
conversion instead; this scoring rule gives those choices equal weight,
which seems to work best in practice.

Unfortunately, this significantly increases the number of candidate
patterns that are tested, and the result is about 90 times slower. But I
can get the performance back with some algorithmic fixes later, now that
I have a reference implementation which seems to produce good results.