Prune according to ordering constraints (super fast!)
I have a few rules which seem to hold true in all modern date and time
formats, like "minutes should come immediately after hours", and "month
and day-of-month should be next to each other, in either order".

Previously I checked these rules at validation-time, after generating a
complete candidate format string. However, if a prefix of the format
string violates some rule, then no matter how you complete the format
string, it will still violate that rule.

So this patch prunes subtrees of the search space as soon as a rule can
be determined to be violated. The result is it runs in 1/16 the time of
the previous commit.

This is now an order of magnitude faster than when I started this
morning, and produces better results too. Hooray!