More precise heuristic for branch-and-bound
I guessed that evaluating the admissible heuristic would be
time-consuming, so I precomputed it. But now I have better timing
measurements, so I tried a more precise heuristic that costs more CPU
time.

It looks like precision yields a 20% speedup for my slowest test cases,
and a maybe 5% slowdown for the fastest ones. (The latter is on the
order of a hundred microseconds, and my naive timing setup isn't
accurate enough to say much at that scale.)

This makes sense because getting a tighter bound in this heuristic can
produce an exponential reduction in the size of the search space, but
adds a constant-factor cost per node that's still explored. The fast
test cases already have small search spaces, so the constant factors
dominate; while the slow cases are slow due to exploring large search
spaces, so exponential reductions there have a significant effect.