Value-ordering by descending optimistic score
This finds high-scoring options sooner so branch-and-bound can prune
more aggressively.

It also gives a small constant-factor speedup in the admissable
heuristic because the maximum heuristic value is always the first option
which satisfies the current constraints, so we don't have to evaluate
all of them to find the maximum.

In practice, this seems to give the best improvement on inputs with
small search spaces. In my larger test cases, I think greedy early
selection is tending to make later options unsatisfiable, but the
constant factor improvements seem to be almost exactly canceling out the
cost of exploring a larger tree.

I previously tried implementing a dynamic value ordering using the same
optimistic-score heuristic, but that didn't improve performance. Since
the optimistic score is actually static, this version can avoid paying
the extra memory allocation and sorting time at every node and just do
it once up-front, which is enough to actually help.