Allow arbitrary variable ordering (slower!)
It's generally better to assign variables that have few options early so
that the harder choices are closer to the leaves of the search tree.

However, my evaluation function was able to aggressively prune using the
knowledge that positions were assigned values in sequential order. So
this version doesn't prune as much of the tree, and therefore takes
longer--not much longer for easy instances, but an order of magnitude
longer for hard instances.