Assign positions to formats, not vice versa
Initially I wrote this search so that, for each position in the input
string, it'd try to pick a good format string to put in that position if
possible.

This commit flips that around: For each category of format string that
can appear in the output, it tries to pick a location to convert in that
format.

This captures the intuition that if only one part of the input could
possibly represent a month, say, then it isn't very likely that we
should interpret that part as anything else.

However this doesn't make performance much better, though I was hoping
it would.