Defer assigning optional fields to late in the search
We want to find valid assignments to the required fields as quickly as
possible. The optional fields can just get skipped if they don't fit
with the earlier choices, but failing a constraint on a required field
prunes however much of the search tree follows it, so we want to fail
those constraints early.