Try positions having few interpretations first
Numeric fields can be interpreted several different ways, while
locale-specific strings usually have only one interpretation.

When we're picking which position to interpret in a particular way, if
we pick one that can't be interpreted in any other ways, then we don't
rule out any options for any of the other assignments we still have to
do.

That increases the chances that we'll find a valid leaf without reaching
a contradiction along the way.

This generalizes: assigning a position which has two interpretations is
less constraining than one which has three, etc.