Factor out multi-variable constraints to functions
I think this makes it a little easier to understand what rules I'm
applying to filter out interpretations that I believe are nonsensical
across all locales. That could be particularly important if I've guessed
incorrectly...

Also this opens up the possibility of using metadata about constraints
to optimize the search.